Changing resource path and adding exception handling for delete operation

+++ src/main/java/br/com/klauddius/filmeflix/movie/controller/MovieController.java
@@ -1,6 +1,7 @@
 package br.com.klauddius.filmeflix.movie.controller;
 
 import br.com.klauddius.filmeflix.movie.entity.Movie;
+import br.com.klauddius.filmeflix.movie.exception.MovieNotFoundException;
 import br.com.klauddius.filmeflix.movie.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -8,6 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class MovieController {
@@ -15,25 +17,29 @@ public class MovieController {
     @Autowired
     private MovieRepository repository;
 
-    @GetMapping("/all")
+    @GetMapping("/movie/all")
     public List<Movie> getAllMovies() {
         return repository.findAll();
     }
 
-    @GetMapping("/all/recently")
+    @GetMapping("/movie/all/recently")
     public List<Movie> getAllMoviesRecentlyAdded() {
         return repository.findAll(Sort.by(Sort.Direction.DESC, "creationDate"));
     }
 
-    @PostMapping("/add")
+    @PostMapping("/movie/add")
     @ResponseStatus(HttpStatus.CREATED)
     public void addMovie(@RequestBody Movie movie) {
         repository.save(movie);
     }
 
-    @DeleteMapping("/delete/{id}")
+    @DeleteMapping("/movie/delete/{id}")
     @ResponseStatus(HttpStatus.ACCEPTED)
     public void deleteMovie(@PathVariable Long id) {
+        Optional<Movie> movieToExclude = repository.findById(id);
+        if (!movieToExclude.isPresent()) {
+            throw new MovieNotFoundException();
+        }
         repository.deleteById(id);
     }
